feat(dresed-react): createMessage, editMessage
